requirements.txt
pandas==1.1.5
PyYAML==5.4.1
numpy==1.16.2
scikit-learn==0.20.3
requests==2.25.1
pytest==6.2.4